Look for fenced code blocks before other block-level constructs
A fenced code block could contain any sort of other block-level construct, for
example:

```
---
foo
---
```

Previously this was interpreted as a two-line header, per GitHub #10.

I also tightened up the regex around looking for a language name on a fenced
code block. This was using "\s" when it should only look for horizontal
whitespace.